+++ java/org/ohdsi/databaseConnector/BatchedQuery.java
@@ -4,6 +4,7 @@ import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
+import java.sql.SQLFeatureNotSupportedException;
 import java.sql.Statement;
 import java.sql.Timestamp;
 import java.sql.Types;
@@ -53,10 +54,18 @@ public class BatchedQuery {
 		rowCount = 0;
 	}
 
+	private void trySettingAutoCommit(Connection connection, boolean value) throws SQLException {
+		try {
+			connection.setAutoCommit(value);
+		} catch (SQLFeatureNotSupportedException exception) {
+			// Do nothing
+		}
+	}
+
 	public BatchedQuery(Connection connection, String query) throws SQLException {
 		this.connection = connection;
 		if (connection.getAutoCommit())
-			connection.setAutoCommit(false);
+			trySettingAutoCommit(connection, false);
 		Statement statement = connection.createStatement(ResultSet.TYPE_FORWARD_ONLY, ResultSet.CONCUR_READ_ONLY);
 		statement.setFetchSize(FETCH_SIZE);
 		resultSet = statement.executeQuery(query);
@@ -107,13 +116,13 @@ public class BatchedQuery {
 						((String[]) columns[columnIndex])[rowCount] = null;
 					else
 						((String[]) columns[columnIndex])[rowCount] = DATETIME_FORMAT.format(timestamp);
-					
+
 				}
 			rowCount++;
 		}
 		if (rowCount < batchSize) {
 			done = true;
-			connection.setAutoCommit(true);
+			trySettingAutoCommit(connection, true);
 		}
 		totalRowCount += rowCount;
 	}